requirements.txt
PyQt5>=5.15.6
Pillow>=10.0.0